Avoid implicit conversion from floats in dart:ui
Dart has a built in `double` type. We currently sometimes let
Dart's FFI mechanisms automatically truncate this to float (like in
the path API), and sometimes implicitly convert Dart doubles to
C++ floats (aka SkScalar and impeller:Scalar). We also do implicit
conversions when dealing with `Float64List` data, especially w.r.t.
matrices. This causes problems when applications use very large
double values which then get silently truncated to infinity, especially
in matrices and values that end up being used in matrix concatenation.

This is solved by the following.

  - Enable `-Wimplicit-float-conversion` for dart:ui's C++
    sources.
  - Add a bunch of `f`s to C++ literals.
  - Create a utility method for truncating doubles to floats
    in a well defined manner, avoiding truncation from a
    finite double value to an infinite float value.
  - Delete Tonic's converters from double to float, since
    these silently truncate to infinity before the engine C++ code
    even touches them.

Fixes https://github.com/flutter/flutter/issues/121299

+++ lib/ui/painting/image_decoder_impeller.cc
@@ -124,8 +124,8 @@ std::shared_ptr<SkBitmap> ImageDecoderImpeller::DecompressTexture(
   auto decode_size = source_size;
   if (descriptor->is_compressed()) {
     decode_size = descriptor->get_scaled_dimensions(std::max(
-        static_cast<double>(target_size.width()) / source_size.width(),
-        static_cast<double>(target_size.height()) / source_size.height()));
+        static_cast<float>(target_size.width()) / source_size.width(),
+        static_cast<float>(target_size.height()) / source_size.height()));
   }
 
   //----------------------------------------------------------------------------
